requirements.txt
pyserial
pygame
matplotlib
numpy